dep: upgrade go to 1.21
