+++ internal/api/medias/controller.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"net/http"
 	"strconv"
-	"time"
 
 	"github.com/go-playground/validator/v10"
 	"github.com/google/uuid"
@@ -15,106 +14,18 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
-// DTOs
-type (
-	// watchTargetDto represents a way in which a particular type
-	// of media could be watched:
-	// - If a particular target has a pre-transcode COMPLETE for the given media,
-	//   then a target to that effect will be present (with ready = true),
-	// - If a particular target has a pre-transcode IN PROGRESS for the given media,
-	//   then a target with ready=false will be present,
-	// - If no pre-transcode for the given target and media is found, AND the target is
-	//   eligible for on-the-fly transcode, then a target of type 'on-the-fly' will
-	//   be present (with ready = true).
-	// - ADDITIONALLY, a watch target with NO target_id will be available of type
-	//   on-the-fly and ready=true will be present, which represents being able to stream
-	//   from the source media directly.
-	watchTargetType string
-	watchTargetDto  struct {
-		Name     string          `json:"display_name"`
-		TargetID *uuid.UUID      `json:"target_id,omitempty"`
-		Enabled  bool            `json:"enabled"`
-		Type     watchTargetType `json:"type"`
-		Ready    bool            `json:"ready"`
-		// TODO: may want to include some additional information about the
-		// target here, such as bitrate and resolution.
-	}
-
-	seriesStubDto struct {
-		Id          uuid.UUID `json:"id"`
-		Title       string    `json:"title"`
-		SeasonCount int       `json:"season_count"`
-	}
-
-	movieStubDto struct {
-		Id    uuid.UUID `json:"id"`
-		Title string    `json:"title"`
-		// TODO: poster path, runtime
-	}
-
-	listDto struct {
-		Type      string    `json:"type"`
-		Id        uuid.UUID `json:"id"`
-		Title     string    `json:"title"`
-		TmdbID    string    `json:"tmdb_id"`
-		CreatedAt time.Time `json:"created_at"`
-		UpdatedAt time.Time `json:"updated_at"`
-		// TODO poster path, optional movie/series specific information such as runtime and season count
-	}
-
-	episodeDto struct {
-		Id           uuid.UUID         `json:"id"`
-		TmdbId       string            `json:"tmdb_id"`
-		Title        string            `json:"title"`
-		CreatedAt    time.Time         `json:"created_at"`
-		UpdatedAt    time.Time         `json:"updated_at"`
-		WatchTargets []*watchTargetDto `json:"watch_targets"`
-	}
-
-	episodeStubDto struct {
-		Id    uuid.UUID `json:"id"`
-		Title string    `json:"title"`
-		Adult bool      `json:"adult"`
-	}
-
-	seasonDto struct {
-		Episodes []*episodeStubDto `json:"episodes"`
-	}
-
-	seriesDto struct {
-		Id      uuid.UUID    `json:"id"`
-		TmdbId  string       `json:"tmdb_id"`
-		Title   string       `json:"title"`
-		Seasons []*seasonDto `json:"seasons"`
-	}
-
-	// movieDto is a fully inflated version of the more common movieStubDto, which encodes more
-	// information such as the watch targets which are eligible for the media
-	movieDto struct {
-		Id           uuid.UUID         `json:"id"`
-		TmdbId       string            `json:"tmdb_id"`
-		Title        string            `json:"title"`
-		CreatedAt    time.Time         `json:"created_at"`
-		UpdatedAt    time.Time         `json:"updated_at"`
-		WatchTargets []*watchTargetDto `json:"watch_targets"`
-	}
-)
-
-const (
-	PreTranscoded watchTargetType = "pre_transcode"
-	LiveTranscode watchTargetType = "live_transcode"
-)
-
 type (
 	Store interface {
 		GetMovie(movieID uuid.UUID) (*media.Movie, error)
 		GetEpisode(episodeID uuid.UUID) (*media.Episode, error)
-		ListMovie() ([]*media.Movie, error)
-		ListSeriesStubs() ([]*media.SeriesStub, error)
-		ListLatestMedia(allowedTypes []string, limit int) ([]*media.Container, error)
 		GetInflatedSeries(seriesID uuid.UUID) (*media.InflatedSeries, error)
 		GetTranscodesForMedia(uuid.UUID) ([]*transcode.Transcode, error)
 		GetAllTargets() []*ffmpeg.Target
+
+		ListLatestMedia(allowedTypes []string, limit int) ([]*media.Container, error)
+		ListMovie() ([]*media.Movie, error)
+		ListSeriesStubs() ([]*media.SeriesStub, error)
+
 		DeleteEpisode(episodeID uuid.UUID) error
 		DeleteSeries(seriesID uuid.UUID) error
 		DeleteSeason(seasonID uuid.UUID) error
@@ -162,7 +73,7 @@ func (controller *Controller) listLatest(ec echo.Context) error {
 	params := ec.QueryParams()
 	allowedTypes, ok := params["allowedType"]
 	if !ok {
-		allowedTypes = []string{"movie", "series"}
+		allowedTypes = []string{}
 	}
 
 	limit, err := strconv.Atoi(params.Get("limit"))
@@ -235,8 +146,8 @@ func (controller *Controller) getMovie(ec echo.Context) error {
 	}
 
 	dto := movieDto{
-		Id:           movie.ID,
-		TmdbId:       movie.TmdbId,
+		ID:           movie.ID,
+		TmdbID:       movie.TmdbID,
 		Title:        movie.Title,
 		CreatedAt:    movie.CreatedAt,
 		UpdatedAt:    movie.UpdatedAt,
@@ -265,8 +176,8 @@ func (controller *Controller) getEpisode(ec echo.Context) error {
 	}
 
 	dto := episodeDto{
-		Id:           episode.ID,
-		TmdbId:       episode.TmdbId,
+		ID:           episode.ID,
+		TmdbID:       episode.TmdbID,
 		Title:        episode.Title,
 		CreatedAt:    episode.CreatedAt,
 		UpdatedAt:    episode.UpdatedAt,
@@ -389,51 +300,6 @@ func (controller *Controller) getMediaWatchTargets(mediaID uuid.UUID) ([]*watchT
 	watchTargets = append(watchTargets, &watchTargetDto{Name: "Source", Ready: true, Type: LiveTranscode, TargetID: nil, Enabled: true})
 
 	return watchTargets, nil
-}
-
-func newWatchTarget(target *ffmpeg.Target, t watchTargetType, ready bool) *watchTargetDto {
-	return &watchTargetDto{Name: target.Label, Ready: ready, Type: t, TargetID: &target.ID, Enabled: true}
-}
-
-func newListDtos(results []*media.Container) ([]*listDto, error) {
-	dtos := make([]*listDto, len(results))
-	for k, v := range results {
-		dto, err := newListDto(v)
-		if err != nil {
-			return nil, err
-		}
-		dtos[k] = dto
-	}
-
-	return dtos, nil
-}
-
-func newListDto(container *media.Container) (*listDto, error) {
-	var ty string
-	if container.Type == media.SERIES {
-		ty = "series"
-	} else if container.Type == media.MOVIE {
-		ty = "movie"
-	} else {
-		return nil, echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Media %s found during listing has an illegal type %s. Expected movie or series.", container, container.Type))
-	}
-
-	return &listDto{Type: ty, Id: container.Id(), Title: container.Title(), TmdbID: container.TmdbId(), CreatedAt: container.CreatedAt(), UpdatedAt: container.UpdatedAt()}, nil
-}
-
-func inflatedSeriesModelToDto(model *media.SeriesStub) seriesStubDto {
-	return seriesStubDto{
-		Id:          model.ID,
-		Title:       model.Title,
-		SeasonCount: model.SeasonCount,
-	}
-}
-
-func movieModelToDto(model *media.Movie) movieStubDto {
-	return movieStubDto{
-		Id:    model.ID,
-		Title: model.Title,
-	}
 }
 
 func wrapErrorGenerator(message string) func(err error) error {
