+++ internal/http/tmdb/media_conv.go
@@ -7,7 +7,7 @@ import (
 
 func TmdbEpisodeToMedia(ep *Episode, isSeasonAdult bool, metadata *media.FileMediaMetadata) *media.Episode {
 	return &media.Episode{
-		Model: media.Model{ID: uuid.New(), TmdbId: ep.Id.String(), Title: ep.Name},
+		Model: media.Model{ID: uuid.New(), TmdbID: ep.Id.String(), Title: ep.Name},
 		Watchable: media.Watchable{
 			MediaResolution: media.MediaResolution{Width: *metadata.FrameW, Height: *metadata.FrameH},
 			SourcePath:      metadata.Path,
@@ -19,19 +19,19 @@ func TmdbEpisodeToMedia(ep *Episode, isSeasonAdult bool, metadata *media.FileMed
 
 func TmdbSeriesToMedia(series *Series) *media.Series {
 	return &media.Series{
-		Model: media.Model{ID: uuid.New(), TmdbId: series.Id.String(), Title: series.Name},
+		Model: media.Model{ID: uuid.New(), TmdbID: series.Id.String(), Title: series.Name},
 	}
 }
 
 func TmdbSeasonToMedia(season *Season) *media.Season {
 	return &media.Season{
-		Model: media.Model{ID: uuid.New(), TmdbId: season.Id.String(), Title: season.Name},
+		Model: media.Model{ID: uuid.New(), TmdbID: season.Id.String(), Title: season.Name},
 	}
 }
 
 func TmdbMovieToMedia(movie *Movie, metadata *media.FileMediaMetadata) *media.Movie {
 	return &media.Movie{
-		Model: media.Model{ID: uuid.New(), TmdbId: movie.Id.String(), Title: movie.Name},
+		Model: media.Model{ID: uuid.New(), TmdbID: movie.Id.String(), Title: movie.Name},
 		Watchable: media.Watchable{
 			MediaResolution: media.MediaResolution{Width: *metadata.FrameW, Height: *metadata.FrameH},
 			SourcePath:      metadata.Path,
